refactor: Add user-action.mock.ts and db.mock.ts files
